refactoring of family scheme away from working for families
