+++ openfisca_aotearoa/tests/family_scheme.yaml
@@ -14,7 +14,7 @@
     - id: "Earth"
       owners: ["Papatūānuku", "Rakinui"]
   output_variables:
-    income_tax__caregiver_age_qualifies_under_family_scheme:
+    family_scheme__caregiver_age_qualifies:
       2018-07: [0, 1]
 - name: Tests if family has dependent children in relation to the family scheme
   period: 2018
@@ -32,5 +32,5 @@
     - id: "Earth"
       owners: ["Papatūānuku", "Hemi"]
   output_variables:
-    income_tax__family_has_dependent_children:
+    family_scheme__has_dependent_children:
       2018-07: [1]
