+++ openfisca_aotearoa/tests/family_scheme_eligibility.yaml
@@ -1,4 +1,4 @@
-# We can run this test on our command line using `openfisca-run-test openfisca_aotearoa/tests/working_for_families.yaml`
+# We can run this test on our command line using `openfisca-run-test openfisca_aotearoa/tests/family_scheme_eligibility.yaml`
 - name: A Person receiving an income tested benefit is not eligible for working for families
   period: 2018-10
   absolute_error_margin: 1
@@ -8,7 +8,7 @@
     veterans_support__received_parents_allowance: False
     veterans_support__received_childrens_pension: False
   output_variables:
-    income_tax__eligible_for_working_for_families: False
+    family_scheme__qualifies_for_entitlements: False
 - name: A Person receiving a parents allowance is not eligible for working for families
   period: 2018-10
   absolute_error_margin: 1
@@ -18,7 +18,7 @@
     veterans_support__received_parents_allowance: True
     veterans_support__received_childrens_pension: False
   output_variables:
-    income_tax__eligible_for_working_for_families: False
+    family_scheme__qualifies_for_entitlements: False
 - name: A Person receiving a childrens pension is not eligible for working for families
   period: 2018-10
   absolute_error_margin: 1
@@ -28,7 +28,7 @@
     veterans_support__received_parents_allowance: False
     veterans_support__received_childrens_pension: True
   output_variables:
-    income_tax__eligible_for_working_for_families: False
+    family_scheme__qualifies_for_entitlements: False
 - name: A Person is eligible for working for families
   period: 2018-10
   absolute_error_margin: 1
@@ -38,7 +38,7 @@
     veterans_support__received_parents_allowance: False
     veterans_support__received_childrens_pension: False
   output_variables:
-    income_tax__eligible_for_working_for_families: True
+    family_scheme__qualifies_for_entitlements: True
 - name: A Person is to young for working for families
   period: 2018-10
   absolute_error_margin: 1
@@ -48,4 +48,4 @@
     veterans_support__received_parents_allowance: False
     veterans_support__received_childrens_pension: False
   output_variables:
-    income_tax__eligible_for_working_for_families: False
+    family_scheme__qualifies_for_entitlements: False
